test: stop shipping the git_repo fixture in the tests bundle
The tests bundle shipped a self-contained, self-testing trio to every
consumer repo — the git_repo fixture (in conftest.py), its only consumer
test_git_repo_fixture.py, and test_utils.py (whose GIT export only the
fixture used). None of it is usable by consumers: it exists purely to
test itself, so it lands as dead code in every downstream project (e.g.
jquantstats had to strip it after deleting the fixture test).

Relocate the git_repo fixture and its self-test into rhiza's own tests/
suite so rhiza keeps exercising the fixture, and remove them from the
bundle:

- bundles/.../conftest.py: keep only the root + logger fixtures
- bundles/.../test_utils.py: deleted (helpers already exist as
  tests/util.py for rhiza's own suite; unused inside the bundle)
- bundles/.../test_git_repo_fixture.py -> tests/test_git_repo_fixture.py
- tests/conftest.py: gains the git_repo fixture + mock uv/make scripts
- drop the now-dangling .rhiza/tests symlinks and update the bundle README

Consumers now receive a lean bundle (pyproject/readme/docstring tests
plus root/logger fixtures). Verified: relocated fixture test + bundle
suite pass (45 passed, 2 skipped); full tests/ tree collects cleanly
(1341 tests).
