--- .rhiza/tests/test_git_repo_fixture.py
@@ -1,1 +1,0 @@
-../../bundles/tests/.rhiza/tests/test_git_repo_fixture.py
