--- .rhiza/tests/test_utils.py
@@ -1,1 +1,0 @@
-../../bundles/tests/.rhiza/tests/test_utils.py
